% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cents_fallpctcover}
\alias{cents_fallpctcover}
\title{Fall cover as percentages of various weeds, cover crops, soil}
\format{
A data frame with 10800 rows and 6 variables
\describe{
\item{eu_id}{A unique identifier for a block, straw trt, till trt, and cover crop (cc) trt}
\item{date2}{The date the pictures were taken (yyyy-mm-dd)}
\item{subrep}{The sub-sample identifer (1, 2, 3)}
\item{cover_cat}{Categories of vegetation (covercrop, other, soil)}
\item{eppo_code}{The EPPO code for the species (except for soil)}
\item{cover_pct}{The percentage of cover attributed to a given EPPO code (0-100)}
}
}
\usage{
cents_fallpctcover
}
\description{
Percentages were assigned post hoc using pictures taken in the field.
The variables are as follows:
}
\keyword{datasets}
