% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cents_herbops}
\alias{cents_herbops}
\title{Data from Bo on herbicide usage, needs confirmed}
\format{
A data frame with 52 rows and 6 variables
\describe{
\item{date2}{The date of the application (yyyy-mm-dd)}
\item{till_id}{Tillage treatment id (notill, noninversion, inversion)}
\item{cctrt_id}{Cover crop treatment id (mix_E, mix_M, rad_M, rad_L, nocc)}
\item{product}{Commerical herbicide product name}
\item{amount}{Amount of herbicide applied per unit area}
\item{amount_units}{Units of amount (liters per hectare, grams per hectare)}
}
}
\usage{
cents_herbops
}
\description{
The variables are as follows:
}
\keyword{datasets}
