% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cents_pliherbs}
\alias{cents_pliherbs}
\title{Data from Middeldatabasen on herbicides used in CENTS 2018-2020}
\format{
A data frame with 40 rows and 6 variables
\describe{
\item{product}{The product name as reported in the methods and materials}
\item{load_cat}{PLI category (health, env behavior, env impact, total)}
\item{load_area}{The amount of product you have to apply per unit area to achieve a load of 1}
\item{load_area_units}{Units of the product application rate to achieve a load of 1 (liters/hectare, grams/hectare)}
\item{load_unit}{Load assigned to a unit of herbicide}
\item{load_unit_units}{B is the Danish letter for load, load per unit of herbicide (liter or gram)}
}
}
\usage{
cents_pliherbs
}
\description{
The variables are as follows:
}
\keyword{datasets}
