% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cents_spweedcount}
\alias{cents_spweedcount}
\title{Counts of spring weeds per m2 in 2019 (one week after herbicide app in oats)
and in 2020 (one week after second herbicide app in faba bean)}
\format{
A data frame with 2880 rows and 5 variables
\describe{
\item{eu_id}{A unique identifier for a block, straw trt, till trt, and cover crop (cc) trt}
\item{date2}{The date of counting (yyyy-mm-dd)}
\item{subrep}{The sub-sample identifer (1, 2, 3)}
\item{weed_type}{The type of weed with two species identified specifically (Dicot, monocot, cirar, equar)}
\item{count}{The number of weeds counted in a 1 m2 area(0-600))}
}
}
\usage{
cents_spweedcount
}
\description{
The variables are as follows:
}
\keyword{datasets}
